+++ src/main/java/net/minecraft/src/EnumToolMaterial.java
@@ -2,11 +2,11 @@ package net.minecraft.src;
 
 public enum EnumToolMaterial
 {
-    WOOD(0, 59, 2.0F, 0, 15),
-    STONE(1, 131, 4.0F, 1, 5),
-    IRON(2, 250, 6.0F, 2, 14),
-    EMERALD(3, 1561, 8.0F, 3, 10),
-    GOLD(0, 32, 12.0F, 0, 22);
+    WOOD(0, 59, 2.0F, 0.0F, 15),
+    STONE(1, 131, 4.0F, 1.0F, 5),
+    IRON(2, 250, 6.0F, 2.0F, 14),
+    EMERALD(3, 1561, 8.0F, 3.0F, 10),
+    GOLD(0, 32, 12.0F, 0.0F, 22);
 
     /**
      * The level of material this tool can harvest (3 = DIAMOND, 2 = IRON, 1 = STONE, 0 = IRON/GOLD)
@@ -24,12 +24,12 @@ public enum EnumToolMaterial
     private final float efficiencyOnProperMaterial;
 
     /** Damage versus entities. */
-    private final int damageVsEntity;
+    private final float damageVsEntity;
 
     /** Defines the natural enchantability factor of the material. */
     private final int enchantability;
 
-    private EnumToolMaterial(int par3, int par4, float par5, int par6, int par7)
+    private EnumToolMaterial(int par3, int par4, float par5, float par6, int par7)
     {
         this.harvestLevel = par3;
         this.maxUses = par4;
@@ -57,7 +57,7 @@ public enum EnumToolMaterial
     /**
      * Returns the damage against a given entity.
      */
-    public int getDamageVsEntity()
+    public float getDamageVsEntity()
     {
         return this.damageVsEntity;
     }
