+++ src/main/java/net/minecraft/src/IEntityMultiPart.java
@@ -4,5 +4,5 @@ public interface IEntityMultiPart
 {
     World func_82194_d();
 
-    boolean attackEntityFromPart(EntityDragonPart var1, DamageSource var2, int var3);
+    boolean attackEntityFromPart(EntityDragonPart var1, DamageSource var2, float var3);
 }
